Test_system/src/auth/requirements.txt
blinker==1.8.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.8
dnspython==2.6.1
flask==3.0.3
Flask-PyMongo==2.3.0
importlib-metadata==8.0.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
pycparser==2.22
PyJWT==2.8.0
pymongo==4.8.0
werkzeug==3.0.3
zipp==3.19.2
